+++ src/sdk/CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Admin library: Group connected team sites use the Microsoft Graph api for both application and delegated permissions #1220 [danielpastoor - Daniel Pastoor]
 - Added Poland to the `GeoLocation` enum #1232 [jansenbe - Bert Jansen]
 - Fix issue to enumerate private/shared channel sites #1238 [valpvt]
+- Removed `ExcludedBlockDownloadGroupIds` properties from `ISiteCollectionProperties` as that prevented unlocking a site collection via the `LockState` property #1251 [jansenbe - Bert Jansen]
 
 ## [1.10.0]